pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cpi-vae"
version = "0.1.0"
description = "Convolutional VAE for CPI cloud particle imagery"
authors = [{name = "Joseph Ko", email = ""}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "torch>=1.12",
    "torchvision",
    "numpy",
    "Pillow",
    "tqdm",
]

[project.optional-dependencies]
dev = ["pytest", "black", "flake8"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
